sensibly combine tests into suites
